requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
openai==1.93.0
streamlit==1.46.1
tiktoken==0.9.0
python-dotenv==1.1.1
graphviz==0.21
